Move tests out of desktop folder
